+++ kohii-exoplayer/src/main/java/kohii/v1/exoplayer/DefaultMediaSourceFactoryProvider.kt
@@ -29,10 +29,8 @@ import com.google.android.exoplayer2.source.smoothstreaming.SsMediaSource
 import com.google.android.exoplayer2.upstream.DataSource
 import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory
 import com.google.android.exoplayer2.upstream.DefaultHttpDataSourceFactory
-import com.google.android.exoplayer2.upstream.FileDataSource
 import com.google.android.exoplayer2.upstream.HttpDataSource
 import com.google.android.exoplayer2.upstream.cache.Cache
-import com.google.android.exoplayer2.upstream.cache.CacheDataSinkFactory
 import com.google.android.exoplayer2.upstream.cache.CacheDataSource
 import com.google.android.exoplayer2.upstream.cache.CacheDataSourceFactory
 import com.google.android.exoplayer2.util.Util
@@ -50,10 +48,6 @@ class DefaultMediaSourceFactoryProvider @JvmOverloads constructor(
   mediaCache: Cache? = null
 ) : MediaSourceFactoryProvider {
 
-  companion object {
-    private const val CACHE_FRAGMENT_SIZE = 2 * 1024 * 1024L /* 2 Megabyte */
-  }
-
   constructor(context: Context, dataSourceFactory: HttpDataSource.Factory) : this(
       dataSourceFactory = DefaultDataSourceFactory(context, dataSourceFactory),
       drmSessionManagerProvider = DefaultDrmSessionManagerProvider(context, dataSourceFactory),
@@ -69,9 +63,7 @@ class DefaultMediaSourceFactoryProvider @JvmOverloads constructor(
     CacheDataSourceFactory(
         mediaCache,
         dataSourceFactory,
-        FileDataSource.Factory(),
-        CacheDataSinkFactory(mediaCache, CACHE_FRAGMENT_SIZE),
-        CacheDataSource.FLAG_IGNORE_CACHE_ON_ERROR, null
+        CacheDataSource.FLAG_IGNORE_CACHE_ON_ERROR
     )
   } else {
     dataSourceFactory